requirements.txt
numpy==1.24.2
pygame==2.1.2
PyYAML==6.0
scipy==1.10.0